Moved From Cross Platform Testing into Master
Current Issues are the setup window does not automatically switch to the main program after setup is complete. As of right now, you have to close the setup and then re open the program. Also, the relaxing auto paint feature on the Games page does not open when using the exe file.